requirements.txt

lightning-bolts
matplotlib
numpy
scipy
scikit-learn
scprep
scanpy
torchdyn>=1.0.6   # 1.0.4 is broken on pypi
pot
torchdiffeq
absl-py
clean-fid
